Add `turbo:before-stream-render` event handler to Intercom
Prior to this change, we'd get a bunch of `NS_ERROR_NOT_INITIALIZED` errors in Firefox browsers when turbo streams updated the DOM.

In Safari, the Intercom widget would disappear, but we'd get no errors.

This change makes Intercom aware of Turbo stream updates, effectively causing it to reinitialise. Ideally the widget wouldn't need to reinitialise at all (as this causes it to disappear and reappear in Safari - Firefox doesn't show the change), but this is better than a bunch of exceptions whenever the user scrolls their mouse.